user investment API working

+++ src/main/java/com/akshat14714/stockcalculator/common/dtos/Stocks.java
@@ -11,7 +11,7 @@ public class Stocks {
 
     private int esg;
 
-    private int stockReturn;
+    private double stockReturn;
 
     private int incentive;
